+++ src/core/client/admin/routes/Configure/sections/General/AnnouncementFormModal.css
@@ -1,0 +1,15 @@
+$announcement-modal-text: var(--v2-colors-mono-500);
+
+.root {
+  min-width: 500px;
+  padding: var(--v2-spacing-2) var(--v2-spacing-3) var(--v2-spacing-3)
+    var(--v2-spacing-3);
+}
+.title {
+  font-size: var(--v2-font-size-5);
+  font-family: var(--v2-font-family-primary);
+  font-weight: var(--v2-font-weight-primary-semi-bold);
+  line-height: var(--v2-line-height-title);
+  color: $announcement-modal-text;
+  margin: 0;
+}
